# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ASKO API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/user-devices/{id}": {
      "get": {
        "operationId": "UserDeviceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeviceRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Devices"
        ]
      },
      "patch": {
        "operationId": "UserDeviceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeviceRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Devices"
        ]
      },
      "delete": {
        "operationId": "UserDeviceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Devices"
        ]
      }
    },
...
    "/dealers": {
      "post": {
        "operationId": "DealerController_createProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDealerProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      },
      "get": {
        "operationId": "DealerController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDealersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
...
    "/dealers/withdrawals/{id}/process": {
      "post": {
        "operationId": "DealerController_processWithdrawal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessWithdrawalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
    "/dealers/withdrawals/{id}/approve": {
      "post": {
        "operationId": "DealerController_approveWithdrawal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
    "/dealers/withdrawals/{id}/reject": {
      "post": {
        "operationId": "DealerController_rejectWithdrawal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
    "/dealers/withdrawals/{id}/mark-paid": {
      "post": {
        "operationId": "DealerController_markPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
    "/dealers/search-user": {
      "get": {
        "operationId": "DealerController_searchUser",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
    "/dealers/profile": {
      "get": {
        "operationId": "DealerController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      },
      "patch": {
        "operationId": "DealerController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
    "/dealers/clients": {
      "get": {
        "operationId": "DealerController_getClients",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerClientListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      },
      "post": {
        "operationId": "DealerController_addClient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDealerClientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerClientResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
    "/dealers/user-devices/{userId}": {
      "get": {
        "operationId": "DealerController_getUserDevices",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerUserDeviceListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
...
    "/dealers/withdraw": {
      "post": {
        "operationId": "DealerController_requestWithdrawal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPointsWithdrawalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dealers"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthProvider": {
        "type": "string",
        "enum": [
          "EMAIL",
          "PHONE",
          "GOOGLE",
          "VK",
          "YANDEX"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "super_admin",
          "admin",
          "dealer",
          "manager",
          "repairer",
          "user"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthProvider"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "providers",
          "roles",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthProvider"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "phoneVerified": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "providers",
          "roles",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "DeviceType": {
        "type": "string",
        "enum": [
          "washing_machine",
          "dryer",
          "dishwasher",
          "oven",
          "cooktop",
          "refrigerator",
          "freezer",
          "hood",
          "other"
        ]
      },
      "DeviceRecordDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DeviceType"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "specifications": {
            "type": "object"
          },
          "features": {
            "type": "object"
          },
          "link": {
            "type": "string"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "categoryId": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "categoryLabel": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "name",
          "model",
          "brand",
          "slug",
          "createdAt",
          "updatedAt"
        ]
      },
      "AddressRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "house": {
            "type": "string"
          },
          "building": {
            "type": "string"
          },
          "apartment": {
            "type": "string"
          },
          "entrance": {
            "type": "string"
          },
          "floor": {
            "type": "string"
          },
          "intercom": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "validationStatus": {
            "type": "string"
          },
          "validationError": {
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "city",
          "street",
          "house",
          "isPrimary"
        ]
      },
      "UserDeviceRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "addressId": {
            "type": "string"
          },
          "purchaseDate": {
            "type": "string"
          },
          "warrantyUntil": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "registrationSignature": {
            "type": "string"
          },
          "registrationSignedPayload": {
            "type": "string"
          },
          "validationStatus": {
            "type": "string"
          },
          "validationError": {
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceRecordDto"
          },
          "address": {
            "$ref": "#/components/schemas/AddressRecordDto"
          }
        },
        "required": [
          "id",
          "userId",
          "deviceId",
          "serialNumber",
          "addressId",
          "createdAt"
        ]
      },
...
      "DealerClientRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "dealerId": {
            "type": "string"
          },
          "clientUserId": {
            "type": "string"
          },
          "clientUser": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "dealerId",
          "clientUserId",
          "createdAt"
        ]
      },
      "DealerProfileRecordDto": {
        "type": "object",
        "properties": {
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DealerClientRecordDto"
            }
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "inn": {
            "type": "string"
          },
          "pointsBalance": {
            "type": "number"
          },
          "agreementSignature": {
            "type": "string"
          },
          "agreementSignedPayload": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "pointsBalance",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "paid",
          "partially_refunded",
          "refunded",
          "failed"
        ]
      },
...
      "UpdateUserDeviceDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDealerProfileDto": {
        "type": "object",
        "properties": {}
      },
      "DealerProfileResponseDto": {
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/DealerProfileRecordDto"
          }
        },
        "required": [
          "profile"
        ]
      },
      "PaginatedDealersResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DealerProfileRecordDto"
            }
          },
          "overallCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "overallCount",
          "page",
          "limit"
        ]
      },
      "WithdrawalStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected",
          "completed"
        ]
      },
      "WithdrawalRecordDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WithdrawalStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "dealerId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "requestedAt": {
            "type": "string"
          },
          "processedAt": {
            "type": "string"
          },
          "processedByUserId": {
            "type": "string"
          },
          "cardNumber": {
            "type": "string"
          },
          "cardHolderName": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "id",
          "dealerId",
          "amount",
          "requestedAt"
        ]
      },
...
      "ProcessWithdrawalDto": {
        "type": "object",
        "properties": {}
      },
      "WithdrawalResponseDto": {
        "type": "object",
        "properties": {
          "withdrawal": {
            "$ref": "#/components/schemas/WithdrawalRecordDto"
          }
        },
        "required": [
          "withdrawal"
        ]
      },
      "PayoutResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            ]
          },
          "paymentId": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "paymentId"
        ]
      },
      "UpdateDealerProfileDto": {
        "type": "object",
        "properties": {}
      },
      "DealerClientListResponseDto": {
        "type": "object",
        "properties": {
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DealerClientRecordDto"
            }
          }
        },
        "required": [
          "clients"
        ]
      },
      "DealerUserDeviceRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          },
          "deviceBrand": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "deviceId",
          "serialNumber",
          "deviceName",
          "deviceModel",
          "deviceBrand"
        ]
      },
      "DealerUserDeviceListResponseDto": {
        "type": "object",
        "properties": {
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DealerUserDeviceRecordDto"
            }
          }
        },
        "required": [
          "devices"
        ]
      },
      "AddDealerClientDto": {
        "type": "object",
        "properties": {}
      },
      "DealerClientResponseDto": {
        "type": "object",
        "properties": {
          "client": {
            "$ref": "#/components/schemas/DealerClientRecordDto"
          }
        },
        "required": [
          "client"
        ]
      },
...
      "RequestPointsWithdrawalDto": {
        "type": "object",
        "properties": {}
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    }
  }
}